refactor(react): extract web components to src/web-components directory
