+++ packages/js/package.json
@@ -21,7 +21,8 @@
   "scripts": {
     "start": "webpack-dev-server --inline --config ./config/webpack.dev.config.js",
     "build": "webpack --config ./config/webpack.prod.config.js",
-    "compile": "rm -rf lib && cross-env NODE_ENV=production babel src --out-dir lib --copy-files"
+    "compile": "rm -rf lib && cross-env NODE_ENV=production babel src --out-dir lib --copy-files",
+    "rebuild": "cd ../react/ && yarn compile && cd ../js && yarn build && yarn start"
   },
   "bugs": {
     "url": "https://github.com/Skywatch24/JS-Library/issues"
@@ -29,6 +30,7 @@
   "homepage": "https://github.com/Skywatch24/JS-Library#readme",
   "dependencies": {
     "@skywatch/api": "^1.1.2",
+    "@skywatch/react": "^1.1.3",
     "flv.js": "^1.5.0",
     "video.js": "^7.11.4"
   }
