+++ packages/js/src/index.js
@@ -1,12 +1,10 @@
 import {CoreManager, Constants} from '@skywatch/api';
+import {CameraViewWebComponent} from '@skywatch/react'
+import '@skywatch/react/lib/style/camera-view.css'
 const {SERVER_URL, API_KEY} = Constants;
 
-import {CameraViewWebComponent} from '../../react/lib/index';
-console.log(CameraViewWebComponent);
 customElements.get('camera-view-web-component') ||
   customElements.define('camera-view-web-component', CameraViewWebComponent);
-
-import '../../react/lib/style/camera-view.css';
 
 const Skywatch = {
   initialize: (serverUrl, apiKey) => {
