+++ packages/react/src/web-components/index.js
@@ -1,0 +1,1 @@
+export * from './CameraViewWebComponent'
